backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.2.1
APScheduler==3.10.4
requests==2.32.3
stripe==11.3.0
